# CPLAT-5152: Optimize builder's use of asset reads to avoid unnecessary rebuilds