--- a/example/builder/state_mixin.over_react.g.dart
+++ b/example/builder/state_mixin.over_react.g.dart
@@ -3,7 +3,7 @@
 part of 'state_mixin.dart';
 
 // **************************************************************************
-// OverReactGenerator
+// OverReactBuilder (package:over_react/src/builder.dart)
 // **************************************************************************
 
 abstract class ExampleStateMixinClass implements _$ExampleStateMixinClass {