--- a/lib/src/component/abstract_transition.over_react.g.dart
+++ b/lib/src/component/abstract_transition.over_react.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of over_react.abstract_transition;
+part of 'abstract_transition.dart';
 
 // **************************************************************************
 // OverReactGenerator