--- a/lib/src/component/aria_mixin.over_react.g.dart
+++ b/lib/src/component/aria_mixin.over_react.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of over_react.aria_mixin;
+part of 'aria_mixin.dart';
 
 // **************************************************************************
 // OverReactGenerator