--- a/lib/src/component/prop_mixins.over_react.g.dart
+++ b/lib/src/component/prop_mixins.over_react.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of over_react.prop_mixins;
+part of 'prop_mixins.dart';
 
 // **************************************************************************
 // OverReactGenerator