--- a/lib/src/component/resize_sensor.over_react.g.dart
+++ b/lib/src/component/resize_sensor.over_react.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of over_react.resize_sensor;
+part of 'resize_sensor.dart';
 
 // **************************************************************************
 // OverReactGenerator