--- a/test/over_react/component_declaration/component_type_checking_test/test_a.over_react.g.dart
+++ b/test/over_react/component_declaration/component_type_checking_test/test_a.over_react.g.dart
@@ -1,6 +1,6 @@
 // GENERATED CODE - DO NOT MODIFY BY HAND
 
-part of test_component.test_a;
+part of 'test_a.dart';
 
 // **************************************************************************
 // OverReactGenerator