--- a/test_fixtures/source_files/missing_over_react_g_part/library.dart
+++ b/test_fixtures/source_files/missing_over_react_g_part/library.dart
@@ -1,0 +1,3 @@
+import 'package:over_react/over_react.dart';
+
+part 'part.dart';