--- a/test_fixtures/source_files/missing_over_react_g_part/part.dart
+++ b/test_fixtures/source_files/missing_over_react_g_part/part.dart
@@ -1,0 +1,17 @@
+part of 'library.dart';
+
+@Factory()
+UiFactory<BasicPartOfLibProps> BasicPartOfLib = _$BasicPartOfLib;
+
+@Props()
+class _$BasicPartOfLibProps extends UiProps {
+  String basicProp;
+}
+
+@Component()
+class BasicPartOfLibComponent extends UiComponent<BasicPartOfLibProps> {
+  @override
+  render() {
+    return Dom.div()('foo');
+  }
+}