--- a/test_fixtures/source_files/part_of_basic_library.g.dart
+++ b/test_fixtures/source_files/part_of_basic_library.g.dart
@@ -1,0 +1,6 @@
+// This file should be ignored because of the .g.dart extension.
+
+part of basic.library;
+
+@Factory()
+UiFactory<UiProps> Ignored;