--- a/web/src/demo_components/button.over_react.g.dart
+++ b/web/src/demo_components/button.over_react.g.dart
@@ -3,7 +3,7 @@
 part of 'button.dart';
 
 // **************************************************************************
-// OverReactGenerator
+// OverReactBuilder (package:over_react/src/builder.dart)
 // **************************************************************************
 
 // React component factory implementation.