--- a/web/src/demo_components/list_group.over_react.g.dart
+++ b/web/src/demo_components/list_group.over_react.g.dart
@@ -3,7 +3,7 @@
 part of 'list_group.dart';
 
 // **************************************************************************
-// OverReactGenerator
+// OverReactBuilder (package:over_react/src/builder.dart)
 // **************************************************************************
 
 // React component factory implementation.