--- a/web/src/demo_components/toggle_button_group.over_react.g.dart
+++ b/web/src/demo_components/toggle_button_group.over_react.g.dart
@@ -3,7 +3,7 @@
 part of 'toggle_button_group.dart';
 
 // **************************************************************************
-// OverReactGenerator
+// OverReactBuilder (package:over_react/src/builder.dart)
 // **************************************************************************
 
 // React component factory implementation.