… build.yaml
… lib/src/builder/generator.dart
… pubspec.yaml
… test/over_react/component_declaration/builder_integration_tests/abstract_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/accessor_mixin_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/abstract_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/accessor_mixin_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/component_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/constant_required_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/do_not_generate_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/namespaced_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/private_props_ddc_bug.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/required_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/required_prop_integration_tests.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/stateful_component_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/backwards_compatible/unassigned_prop_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/component_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/constant_required_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/do_not_generate_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/namespaced_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/private_props_ddc_bug.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/required_accessor_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/required_prop_integration_tests.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/stateful_component_integration_test.over_react.g.dart
… test/over_react/component_declaration/builder_integration_tests/unassigned_prop_integration_test.over_react.g.dart
… test/over_react/component_declaration/component_type_checking_test/test_b.over_react.g.dart
… test/over_react/component_declaration/component_type_checking_test/type_inheritance/abstract_inheritance/abstract.over_react.g.dart
… test/over_react/component_declaration/component_type_checking_test/type_inheritance/abstract_inheritance/extendedtype.over_react.g.dart
… test/over_react/component_declaration/component_type_checking_test/type_inheritance/parent.over_react.g.dart
… test/over_react/component_declaration/component_type_checking_test/type_inheritance/subsubtype.over_react.g.dart