# [CPLAT-5152](https://jira.atl.workiva.net/browse/CPLAT-5152)
![Issue Status](http://bender.workiva.org:9000/Bender/status/CPLAT-5152)

## Motivation
  <!-- High-level overview of what you are trying to fix or improve, and why.
         Include any relevant background information that reviewers should know. -->
We've seen unexpectedly large rebuild times in codebases that consume this package and its builder. After some digging, it has been determined that the main reason for this is due to how the over_react builder is currently implemented:

The `PartBuilder` util from the `source_gen` package always retrieves the `LibraryElement` for each file that is a library (i.e. not a part file). This requires that the analysis session resolves the entire element tree for that file, which is expensive but more importantly results in reads on all transitively imported files. These reads are then stored in the asset graph as inputs for that build step, which consequently means that the over_react builder will be re-run on many files unnecessarily because the inputs have changed.

## Changes
  <!-- What this PR changes to fix the problem. -->
The over_react builder only needs the AST for each file, so the solution here is to replace our usage of PartBuilder with a custom Builder that manually reads the library source for each BuildStep and calls the analyzer API `parseCompilationUnit()` to get the info needed to generate the `.over_react.g.dart` part files.

## Perf Improvement

Some numbers to demonstrate the performance improvement delivered by this change:

**Average clean build time:**

> Averaged over 5 runs.
> `pub run build_runner clean && pub run build_runner build`

Project | Before | After
------- | ------ | -----
web_skin_dart | 122s | 107s
wdesk_sdk | 112s | 45s


**Average re-build time:**

> Averaged over 5 runs. Rebuilt after changing one entry point file in `lib/`
> that has a large number of imports/exports.
>
> Change file -> `pub run build_runner build`

Project | File | Before | After
------- | ---- | ------ | -----
web_skin_dart | `lib/ui_core.dart` | 42.9s | 21.22s
wdesk_sdk | `lib/app_infrastructure.dart` | 63s | 2.2s

#### Release Notes
  <!-- A concise description of your changes, written in the imperative.
         ("Fix bug" and not "Fixed bug" or "Fixes bug.") -->
Optimize the over_react builder to avoid unnecessary asset reads. Informal testing in a large codebase that consumes `over_react` has shown in the worst case a speed up on hot-rebuilds from ~1 minute to ~2 seconds.

## Review
_[See CONTRIBUTING.md][contributing-review-types] for more details on review types (+1 / QA +1 / +10) and code review process._

  <!-- If you're making a PR from outside of the Client Platform team, then first off, thanks! :)

        For open-source contributors, tag @Workiva/app-frameworks and we'll take a look!

        For Workiva employees:

            *** Please refrain from tagging the whole team to prevent extraneous notifications. ***

            If you're not sure who from our team should review these changes, then leave this section
            blank for now and post a link to the PR in the #support-ui-platform Slack channel.

  -->

Please review: <!-- Tag people here or via GitHub's "Request Review" feature-->

### QA Checklist
- [ ] Tests were updated and provide good coverage of the changeset and other affected code
- [ ] Manual testing was performed if needed
    - [ ] Steps from PR author: 
        <!-- Call out any specific manual testing instructions here, or omit this section if not applicable -->
    - [ ] Anything falling under manual testing criteria [outlined in CONTRIBUTING.md][contributing-manual-testing]

## Merge Checklist
While we perform many automated checks before auto-merging, some manual checks are needed:
- [ ] A Client Platform member has reviewed these changes
- [ ] There are no unaddressed comments _- this check can be automated if reviewers use the "Request Changes" feature_
- [ ] _For release PRs -_ Version metadata in Rosie comment is correct


[contributing-review-types]: https://github.com/Workiva/over_react/blob/master/CONTRIBUTING.md#review-types
[contributing-manual-testing]: https://github.com/Workiva/over_react/blob/master/CONTRIBUTING.md#manual-testing-criteria