@evanweible-wf 2019-04-02
@greglittlefield-wf @corwinsheahan-wf Just pushed a commit to tidy up the builder implementation a bit.

> Have we let the Dart team know about this performance issue?
Not yet, I want to do a few more tests with a simpler builder and simpler codebase that still uses `buildStep.inputLibrary` and then I plan on asking in their gitter chat about that cost in general. Will probably end up opening an issue in the `build` project around documenting those potential pitfalls as well as an issue in `source_gen` since there is currently no way to opt out of that behavior with their utils.