@evanweible-wf 2019-04-04
Some numbers to demonstrate the performance improvement delivered by this change:

**Average clean build time:**

Project | Before | After
------- | ------ | -----
web_skin_dart | 122s | 107s
wdesk_sdk | 112s | 45s

> Averaged over 5 runs.

---

**Average re-build time:**

Project | File | Before | After
------- | ---- | ------ | -----
web_skin_dart | `lib/ui_core.dart` | 42.9s | 21.22s
wdesk_sdk | `lib/app_infrastructure.dart` | 63s | 2.2s

>  Averaged over 5 runs. Rebuilt after changing one entry point file in `lib/` that has a large number of imports/exports.